Pass PHPStan

+++ TempRanks/src/Heisenburger69/TempRanks/Main.php
@@ -4,38 +4,49 @@ declare(strict_types=1);
 
 namespace Heisenburger69\TempRanks;
 
+use SQLite3;
+use SQLite3Stmt;
+use SQLite3Result;
+use pocketmine\utils\Config;
+use _64FF00\PurePerms\PPGroup;
 use pocketmine\event\Listener;
-use pocketmine\event\player\PlayerJoinEvent;
+use pocketmine\command\Command;
+use _64FF00\PurePerms\PurePerms;
 use pocketmine\plugin\PluginBase;
 use pocketmine\command\CommandSender;
-use pocketmine\command\Command;
-use pocketmine\utils\Config;
 use pocketmine\utils\TextFormat as C;
+use pocketmine\event\player\PlayerJoinEvent;
 
 class Main extends PluginBase implements Listener
 {
-
-    /* @var Config */
-    public $config;
+    public Config $config;
+    public SQLite3 $db;
 
     public function onEnable(): void
     {
         $this->config = new Config($this->getDataFolder() . "config.yml", Config::YAML, ["Check Rank Expiry every 60 seconds" => true, "Time Left Message" => "You have {time_left} on your temporary {temprank} rank", "Rank Expired Message" => "Your {temprank} Rank has expired"]);
         $this->getServer()->getPluginManager()->registerEvents($this, $this);
-        $this->db = new \SQLite3($this->getDataFolder() . "Ranks.db");
+        $this->db = new SQLite3($this->getDataFolder() . "Ranks.db");
         $this->db->exec("CREATE TABLE IF NOT EXISTS ranks (player TEXT PRIMARY KEY COLLATE NOCASE, oldrank TEXT, endtime TEXT);");
         if ($this->config->get("Check Rank Expiry every 60 seconds") === true) {
             $this->getScheduler()->scheduleRepeatingTask(new CheckTask($this), 20 * 60 * 60 * 6);
         }
     }
 
-    public function onJoin(PlayerJoinEvent $event)
+    public function onJoin(PlayerJoinEvent $event): void
     {
+        /** @var PurePerms */
+        $pp = $this->getServer()->getPluginManager()->getPlugin("PurePerms");
+
         $player = $event->getPlayer();
         $playername = $player->getName();
         $time = $this->getTimeLeft($playername);
-        $pp = $this->getServer()->getPluginManager()->getPlugin("PurePerms");
+        
         $rank = $pp->getUserDataMgr()->getGroup($pp->getPlayer($playername));
+        if (!$rank instanceof PPGroup) {
+            return;
+        }
+
         if ($time !== null && $time !== "No temprank") {
             $msg = $this->config->get("Time Left Message");
             $msg = str_replace(array("{time_left}", "{temprank}"), array($time, $rank), $msg);
@@ -47,7 +58,7 @@ class Main extends PluginBase implements Listener
         }
         if (strtotime($exp) < time()) {
             $msg = $this->config->get("Rank Expired Message");
-            $msg = str_replace("{temprank}", $rank, $msg);
+            $msg = str_replace("{temprank}", $rank->getName(), $msg);
             $player->sendMessage($msg);
             $this->removeRank($playername);
         }
@@ -57,40 +68,56 @@ class Main extends PluginBase implements Listener
     {
         switch ($command->getName()) {
             case "temprank":
+                /** @var PurePerms */
                 $pp = $this->getServer()->getPluginManager()->getPlugin("PurePerms");
                 if ($sender->hasPermission("temprank.command")) {
                     if (isset($args[0])) {
                         if ($args[0] === "set") {
                             if (count($args) < 4) {
                                 $sender->sendMessage(C::RED . "Do /temprank set <player> <rank> <duration>\n" . C::AQUA . "For Example: /temprank set Steve Admin 1d12h\nYear = y, Month = m, Day = d, Hour = h, Minute = i, Seconds = s");
-                                return true;
+                                return false;
                             }
                             $oldrank = $pp->getUserDataMgr()->getGroup($pp->getPlayer($args[1]));
                             $player = $this->getServer()->getPlayerByPrefix($args[0]);
                             if ($player !== null) {
                                 $playername = $player->getName();
                             } else {
-                                $playername = (string)$args[1];
+                                $playername = $args[1];
                             }
                             $group = $pp->getGroup($args[2]);
                             if ($group !== null) {
-                                $time = $this->parseTimeFormat($args[3]);
-                                $endtime = date("Y-m-d H:i:s", strtotime(date("Y-m-d H:i:s")) + $time);
+                                $time1 = strtotime(date("Y-m-d H:i:s"));
+                                $time2 = $this->parseTimeFormat($args[3]);
+                                if (!is_int($time2) || !is_int($time1)) {
+                                    return false;
+                                }
+                                $endtime = date("Y-m-d H:i:s", $time1 + $time2);
+
                                 $stmt = $this->db->prepare("INSERT OR REPLACE INTO ranks (player, oldrank, endtime) VALUES (:player, :oldrank, :endtime);");
+                                if (!$stmt instanceof SQLite3Stmt) {
+                                    return false;
+                                }
+
                                 $stmt->bindValue(":player", $playername);
                                 $stmt->bindValue(":oldrank", $oldrank);
                                 $stmt->bindValue(":endtime", $endtime);
                                 $stmt->execute();
                                 $ppplayer = $pp->getPlayer($args[1]);
                                 $pp->setGroup($ppplayer, $group);
-                                $diff = strtotime($endtime) - time();
+
+                                $endTimeTime = strtotime($endtime);
+                                if (!$endTimeTime) {
+                                    return false;
+                                }
+
+                                $diff = $endTimeTime - time();
                                 $length = $this->parseSecondToHuman($diff);
                                 $sender->sendMessage(C::GREEN . "You set the $group Rank to $playername for $length");
                             }
                         } elseif ($args[0] === "remove") {
                             if (count($args) < 2) {
                                 $sender->sendMessage(C::RED . "Use /temprank remove <player>");
-                                return true;
+                                return false;
                             }
                             $playername = $args[1];
                             $this->removeRank($playername);
@@ -107,7 +134,7 @@ class Main extends PluginBase implements Listener
         return true;
     }
 
-    public function getTimeLeft($playername)
+    public function getTimeLeft(string $playername): ?string
     {
         $date = date("Y-m-d H:i:s");
         $enddate = $this->getExpiryDate($playername);
@@ -117,8 +144,13 @@ class Main extends PluginBase implements Listener
         if (strtotime($enddate) < time()) {
             return null;
         }
+
         $datetime1 = date_create($date);
         $datetime2 = date_create($enddate);
+        if (!$datetime1 || !$datetime2) {
+            return null;
+        }
+
         $interval = date_diff($datetime1, $datetime2);
         $min = $interval->format('%i');
         $sec = $interval->format('%s');
@@ -128,7 +160,6 @@ class Main extends PluginBase implements Listener
         $year = $interval->format('%y');
         if ($interval->format('%i%h%d%m%y') == "00000") {
             return $sec . " Seconds";
-
         } else if ($interval->format('%h%d%m%y') == "0000") {
             return $min . " Minutes";
         } else if ($interval->format('%d%m%y') == "000") {
@@ -142,32 +173,49 @@ class Main extends PluginBase implements Listener
         }
     }
 
-    public function getExpiryDate($playername)
+    public function getExpiryDate(string $playername): ?string
     {
         $stmt = $this->db->prepare("SELECT endtime FROM ranks WHERE player = :player;");
+        if (!$stmt instanceof SQLite3Stmt) {
+            return null;
+        }
         $stmt->bindValue(":player", $playername);
+
         $result = $stmt->execute();
+        if (!$result instanceof SQLite3Result) {
+            return null;
+        }
+
         $resultArr = $result->fetchArray(SQLITE3_ASSOC);
-        if (empty($resultArr)) {
+        if (!$resultArr) {
             return null;
         }
         return (string)$resultArr["endtime"];
     }
 
-    public function getOldGroup($playername)
+    public function getOldGroup(string $playername): ?string
     {
         $stmt = $this->db->prepare("SELECT oldrank FROM ranks WHERE player = :player;");
-        $stmt->bindValue(":player", $playername);
-        $result = $stmt->execute();
-        $resultArr = $result->fetchArray(SQLITE3_ASSOC);
-        if (empty($resultArr)) {
+        if (!$stmt instanceof SQLite3Stmt) {
             return null;
         }
-        return (string)$resultArr["oldrank"];
+        $stmt->bindValue(":player", $playername);
+
+        $result = $stmt->execute();
+        if (!$result instanceof SQLite3Result) {
+            return null;
+        }
+
+        $resultArr = $result->fetchArray(SQLITE3_ASSOC);
+        if (!$resultArr) {
+            return null;
+        }
+        return (string) $resultArr["oldrank"];
     }
 
-    public function removeRank($playername)
+    public function removeRank(string $playername): void
     {
+        /** @var PurePerms */
         $pp = $this->getServer()->getPluginManager()->getPlugin("PurePerms");
         $ppplayer = $pp->getPlayer($playername);
         $group = $this->getOldGroup($playername);
@@ -176,14 +224,23 @@ class Main extends PluginBase implements Listener
             $pp->setGroup($ppplayer, $ppgroup);
             return;
         }
+
         $ppgroup = $pp->getGroup($group);
+        if (!$ppgroup instanceof PPGroup) {
+            return;
+        }
         $pp->setGroup($ppplayer, $ppgroup);
+
         $stmt = $this->db->prepare("DELETE FROM ranks WHERE player = :player;");
+        if (!$stmt instanceof SQLite3Stmt) {
+            return;
+        }
+
         $stmt->bindValue(":player", $playername);
         $stmt->execute();
     }
 
-    public function parseTimeFormat(string $duration): ?int//////Thanks Thunder
+    public function parseTimeFormat(string $duration): ?int
     {
         $parts = str_split($duration);
         $time_units = ['y' => 'year', 'm' => 'month', 'w' => 'week', 'd' => 'day', 'h' => 'hour', 'i' => 'minute', 's' => 'second']; //Array of replacement
@@ -207,12 +264,11 @@ class Main extends PluginBase implements Listener
         return $epoch;
     }
 
-    public function parseSecondToHuman($seconds): ?string/////Thanks Thunder
+    public function parseSecondToHuman(int $seconds): ?string
     {
         $dt1 = new \DateTime("@0");
         $dt2 = new \DateTime("@$seconds");
         $diff = $dt1->diff($dt2);
-        if ($diff === false) return null;
         $str = [];
         if ($diff->y > 0) $str[] = $diff->y . ' year(s)';
         if ($diff->m > 0) $str[] = $diff->m . ' month(s)';
@@ -227,6 +283,4 @@ class Main extends PluginBase implements Listener
         }
         return $str;
     }
-
-
 }
